+++ src/main/java/com/rabbitmq/stream/metrics/NoOpMetricsCollector.java
@@ -43,7 +43,7 @@ public final class NoOpMetricsCollector implements MetricsCollector {
     }
 
     @Override
-    public void consume(int count) {
+    public void consume(long count) {
 
     }
 }
